+++ packages/camera/camera_android_camerax/test/camera_test.mocks.dart
@@ -18,6 +18,24 @@ import 'test_camerax_library.g.dart' as _i2;
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
+/// A class which mocks [TestCameraHostApi].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockTestCameraHostApi extends _i1.Mock implements _i2.TestCameraHostApi {
+  MockTestCameraHostApi() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  int getCameraInfo(int? identifier) => (super.noSuchMethod(
+        Invocation.method(
+          #getCameraInfo,
+          [identifier],
+        ),
+        returnValue: 0,
+      ) as int);
+}
+
 /// A class which mocks [TestInstanceManagerHostApi].
 ///
 /// See the documentation for Mockito's code generation for more information.